changed double to int

+++ edu/jalc/watch/interfaces/Settable.java
@@ -1,5 +1,5 @@
 package edu.jalc.watch.interfaces;
 
 public interface Settable{
-  public void set(double time);
+  public void set(int time);
 }
